backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
httpx>=0.26.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
authlib>=1.3.0
itsdangerous>=2.1.2
apscheduler>=3.10.4
pydantic>=2.5.3
pydantic-settings>=2.1.0
requests>=2.31.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
email-validator>=2.0.0
brotli>=1.1.0
